+++ main-service/src/main/java/ru/practicum/main_service/exception/ErrorHandler.java
@@ -1,107 +1,133 @@
 package ru.practicum.main_service.exception;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.MissingServletRequestParameterException;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
-import ru.practicum.main_service.MainCommonUtils;
-
-import javax.validation.ConstraintViolationException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.time.LocalDateTime;
-import java.util.Objects;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import ru.practicum.main_service.exceptions.*;
+import ru.practicum.main_service.exceptions.NotFoundException;
 
 @RestControllerAdvice
 @Slf4j
 public class ErrorHandler {
-    @Getter
-    @AllArgsConstructor
-    private static class ApiError {
-        private final String status;
-        private final String reason;
-        private final String message;
-        private final String errors;
-        private final String timestamp;
-    }
 
-    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ExceptionHandler
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ApiError handleValidationException(final MethodArgumentNotValidException exception) {
-        log.error(exception.toString());
-        return new ApiError(HttpStatus.BAD_REQUEST.name(),
-                "Incorrectly made request.",
-                String.format("Field: %s. Error: %s", Objects.requireNonNull(exception.getFieldError()).getField(),
-                        exception.getFieldError().getDefaultMessage()),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    public ApiError handleMethodArgumentNotValidException(final MethodArgumentNotValidException e) {
+        List<String> stList = Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList());
+        final List<ErrorResponse> errorResponses = e.getBindingResult().getFieldErrors().stream()
+                .map(error -> {
+                    String mes = "Field: " + error.getField() + ". Error: " +
+                            error.getDefaultMessage() + " Value: " + error.getRejectedValue() + " ";
+                    return new ErrorResponse(mes);
+                })
+                .collect(Collectors.toList());
+        StringBuilder stringBuilderErrors = new StringBuilder();
+        for (ErrorResponse errorResponse : errorResponses) {
+            stringBuilderErrors.append(errorResponse.getError());
+        }
+
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.BAD_REQUEST);
+        apiError.setReason("Incorrectly made request.");
+        apiError.setMessage(stringBuilderErrors.toString());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(stList);
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
-    @ExceptionHandler({MethodArgumentTypeMismatchException.class, ConstraintViolationException.class})
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ApiError handleValidationException(final RuntimeException exception) {
-        log.error(exception.toString());
-        return new ApiError(HttpStatus.BAD_REQUEST.name(),
-                "Incorrectly made request.",
-                exception.getMessage(),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.CONFLICT)
+    public ApiError handleNotSave(final NotSaveException e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.CONFLICT);
+        apiError.setReason("Integrity constraint has been violated.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
-    @ExceptionHandler(NotFoundException.class)
+    @ExceptionHandler
     @ResponseStatus(HttpStatus.NOT_FOUND)
-    public ApiError handleNotFoundException(final NotFoundException exception) {
-        log.error(exception.toString());
-        return new ApiError(HttpStatus.NOT_FOUND.name(),
-                "The required object was not found.",
-                exception.getMessage(),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    public ApiError handleNotSave(final NotFoundException e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.NOT_FOUND);
+        apiError.setReason("The required object was not found.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
-    @ExceptionHandler(DataIntegrityViolationException.class)
+    @ExceptionHandler
     @ResponseStatus(HttpStatus.CONFLICT)
-    public ApiError handleDataIntegrityViolationException(final DataIntegrityViolationException exception) {
-        log.error(exception.toString());
-        return new ApiError(HttpStatus.CONFLICT.name(),
-                "Integrity constraint has been violated.",
-                exception.getMessage(),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    public ApiError handleNotSave(final ConflictException e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.CONFLICT);
+        apiError.setReason("For the requested operation the conditions are not met.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
-    @ExceptionHandler(ForbiddenException.class)
-    @ResponseStatus(HttpStatus.CONFLICT)
-    public ApiError handleForbiddenException(final ForbiddenException exception) {
-        log.error(exception.toString());
-        return new ApiError(HttpStatus.CONFLICT.name(),
-                "For the requested operation the conditions are not met.",
-                exception.getMessage(),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiError handleNotSave(final BadRequestException e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.BAD_REQUEST);
+        apiError.setReason("Incorrectly made request.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiError handleMissingRequestParameter(final MissingServletRequestParameterException e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.BAD_REQUEST);
+        apiError.setReason("Required request parameter is not present.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
     @ExceptionHandler
     @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public ApiError handleException(final RuntimeException exception) {
-        log.error("Error 400: {}", exception.getMessage(), exception);
-        return new ApiError(HttpStatus.INTERNAL_SERVER_ERROR.name(),
-                "Unhandled exception.",
-                exception.getMessage(),
-                getErrors(exception),
-                LocalDateTime.now().format(MainCommonUtils.DT_FORMATTER));
+    public ApiError handleThrowable(final Throwable e) {
+        ApiError apiError = new ApiError();
+        apiError.setStatus(HttpStatus.INTERNAL_SERVER_ERROR);
+        apiError.setReason("Произошла непредвиденная ошибка.");
+        apiError.setMessage(e.getMessage());
+        apiError.setTimestamp(LocalDateTime.now());
+        apiError.setErrors(Arrays.stream(e.getStackTrace()).map(x -> x.toString()).collect(Collectors.toList()));
+
+        log.warn(apiError.toString());
+        return apiError;
     }
 
-    private String getErrors(Exception exception) {
-        StringWriter stringWriter = new StringWriter();
-        PrintWriter printWriter = new PrintWriter(stringWriter);
-        exception.printStackTrace(printWriter);
-        return stringWriter.toString();
-    }
 }
